requirements.txt
accelerate==1.2.1
einops==0.8.0
ema-pytorch==0.7.7
gin-config==0.5.0
gymnasium==1.0.0
mujoco==3.2.6
numpy==2.2.1
torch==2.5.1
torchdiffeq==0.2.5
tqdm==4.67.1
wandb==0.19.1
-e git+https://github.com/watchernyu/REDQ.git@3a2763889223677d74c734864244c41494e4832c#egg=redq